+++ ums-service-schedule/src/main/java/com/uth/ums/schedule/model/dto/CourseOccurrenceDto.java
@@ -18,6 +18,6 @@ public class CourseOccurrenceDto implements Serializable {
     String classCode;
     LocalDate startDate;
     LocalDate endDate;
-    int capacity;
+    Long capacity;
     Set<ScheduleDto> schedules;
 }
